% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addch.r
\name{addch}
\alias{addch}
\title{work much like printf()}
\usage{
addch(ch, win = stdscr())
}
\arguments{
\item{ch}{a character to print}

\item{win}{a window object as an external pointer}
}
\description{
If you want to move the cursor first and then print using printw() function,
use move() first and then use printw()
}
\seealso{
Other output: \code{\link{clear}}, \code{\link{mvprintw}},
  \code{\link{printw}}, \code{\link{refresh}},
  \code{\link{waddch}}, \code{\link{wrefresh}}
}

