% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{assume_default_colors}
\alias{assume_default_colors}
\title{tells which colors to paint for color pair 0}
\usage{
assume_default_colors(b, f)
}
\arguments{
\item{b}{the background color number. -1 means the default terminal color

die if the terminal does not support this extension}

\item{f}{the foreground color number. -1 means the default terminal color}
}
\description{
tells which colors to paint for color pair 0
}
\seealso{
Other color: \code{\link{COLOR_PAIR}}, \code{\link{COLOR}},
  \code{\link{PAIR_NUMBER}},
  \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{pair_content}},
  \code{\link{start_color}},
  \code{\link{use_default_colors}}
}

