% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{attron}
\alias{attron}
\title{turns on the given attributes without affecting any others.}
\usage{
attron(...)
}
\arguments{
\item{...}{any valid attributes}
}
\description{
turns on the given attributes without affecting any others.
}
\seealso{
Other attributes: \code{\link{attributes}},
  \code{\link{attroff}}, \code{\link{with_attr}}
}

