% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{box}
\alias{box}
\title{draw a box around a window}
\usage{
box(win = stdscr())
}
\arguments{
\item{win}{a window object as an external pointer}
}
\description{
draw a box around a window
}
\seealso{
Other window: \code{\link{delwin}}, \code{\link{mvwin}},
  \code{\link{newwin}}, \code{\link{stdscr}},
  \code{\link{subwin}}
}

