% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{can_change_color}
\alias{can_change_color}
\title{does the  terminal  supports  colors and can change their definitions}
\usage{
can_change_color()
}
\value{
TRUE iff only it can
}
\description{
Lots of terminals do not support it.
Setting TERM=xterm-256color in terminology on ubuntu seem to enable
support for it (on terminology and gnome-terminal)
}
\seealso{
Other color: \code{\link{COLORS}},
  \code{\link{COLOR_PAIR}}, \code{\link{color_content}},
  \code{\link{color_off}}, \code{\link{color_on}},
  \code{\link{has_colors}}, \code{\link{init_pair}},
  \code{\link{pair_content}}, \code{\link{start_color}}
}

