% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{cbreak}
\alias{cbreak}
\title{disable line buffering and erase/kill character-processing}
\usage{
cbreak(bf = TRUE)
}
\arguments{
\item{bf}{whether to enable or disable}
}
\description{
Normally, the tty driver buffers typed characters until a newline or carriage return is typed.
The cbreak routine disables line buffering and erase/kill character-processing
(interrupt and flow control characters are unaffected), making characters typed by the user
immediately available to the program.
The nocbreak routine  returns the terminal to normal (cooked) mode.
}
\details{
Initially  the terminal may or may not be in cbreak mode, as the mode is inherited; therefore, a program should
       call cbreak or nocbreak explicitly.  Most interactive programs using curses set the  cbreak  mode.
 Note  that
       cbreak  overrides raw.
[See getch(3NCURSES) for a discussion of how these routines interact with echo and noecho.]
}
\seealso{
nocbreak

Other inopts: \code{\link{echo}}, \code{\link{halfdelay}},
  \code{\link{intrflush}}, \code{\link{keypad}},
  \code{\link{meta}}, \code{\link{nocbreak}},
  \code{\link{noecho}}, \code{\link{noqiflush}},
  \code{\link{noraw}}, \code{\link{qiflush}},
  \code{\link{raw}}, \code{\link{timeout}}
}

