% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{clear}
\alias{clear}
\title{clear the screen}
\usage{
clear()
}
\description{
The erase and werase routines copy blanks to every position in the window,
clearing the screen.
The  clear  and wclear routines are like erase and werase,
 but they also call clearok, so that the screen is cleared completely
 on the next call to wrefresh for that window and repainted from scratch.
}
\seealso{
Other output: \code{\link{mvprintw}}, \code{\link{printw}},
  \code{\link{refresh}}
}

