% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{color_content}
\alias{color_content}
\title{get the RGB components of a color}
\usage{
color_content(color)
}
\arguments{
\item{color}{the color index}
}
\value{
the RGB as a named list (R=, G=, B=)
}
\description{
get the RGB components of a color
}
\seealso{
Other color: \code{\link{COLORS}},
  \code{\link{COLOR_PAIR}}, \code{\link{PAIR_NUMBER}},
  \code{\link{can_change_color}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{pair_content}},
  \code{\link{start_color}}
}

