% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{color_off}
\alias{color_off}
\title{unset the current color}
\usage{
color_off(pair)
}
\arguments{
\item{pair}{the pair number}
}
\description{
unset the current color
}
\seealso{
Other color: \code{\link{COLORS}},
  \code{\link{COLOR_PAIR}}, \code{\link{can_change_color}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{start_color}}
}

