% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{color_on}
\alias{color_on}
\title{set the current color}
\usage{
color_on(pair)
}
\arguments{
\item{pair}{the pair number}
}
\description{
set the current color
}
\seealso{
Other color: \code{\link{COLORS}},
  \code{\link{COLOR_PAIR}}, \code{\link{PAIR_NUMBER}},
  \code{\link{assume_default_colors}},
  \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{has_colors}}, \code{\link{init_pair}},
  \code{\link{pair_content}}, \code{\link{start_color}},
  \code{\link{use_default_colors}}
}

