% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.r
\name{COLOR_PAIRS}
\alias{COLOR_PAIRS}
\title{number of color pairs which the terminal can support.}
\usage{
COLOR_PAIRS()
}
\description{
requires initializing curses
Usually the number of color pairs  will  be  the  product COLORS*COLORS,
 however this is not always true:
}
\seealso{
Other variables: \code{\link{COLORS}}, \code{\link{COLS}},
  \code{\link{ESCDELAY}}, \code{\link{LINES}},
  \code{\link{TABSIZE}}, \code{\link{curscr}},
  \code{\link{newscr}}, \code{\link{stdscr}}
}

