% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{COLORS}
\alias{COLORS}
\title{Color constants}
\description{
Color constants
}
\seealso{
keyname

Other color: \code{\link{COLOR_PAIR}},
  \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{start_color}}
}

