% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.r
\name{COLS}
\alias{COLS}
\title{the  width of the screen in characters.}
\usage{
COLS()
}
\value{
the number of columns
}
\description{
requires initializing curses
}
\seealso{
Other variables: \code{\link{COLORS}},
  \code{\link{COLOR_PAIRS}}, \code{\link{ESCDELAY}},
  \code{\link{LINES}}, \code{\link{TABSIZE}},
  \code{\link{curscr}}, \code{\link{newscr}},
  \code{\link{stdscr}}
}

