% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.r
\name{curscr}
\alias{curscr}
\title{special window curscr to record its updates to the terminal screen.}
\usage{
curscr()
}
\value{
curscr as an external pointer
}
\description{
special window curscr to record its updates to the terminal screen.
}
\seealso{
Other variables: \code{\link{COLORS}},
  \code{\link{COLOR_PAIRS}}, \code{\link{COLS}},
  \code{\link{ESCDELAY}}, \code{\link{LINES}},
  \code{\link{TABSIZE}}, \code{\link{newscr}},
  \code{\link{stdscr}}
}

