% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{delwin}
\alias{delwin}
\title{delete the window}
\usage{
delwin(win)
}
\arguments{
\item{win}{a window object as an external pointer}
}
\value{
TRUE iff the window has been deleted
}
\description{
free all memory associated with it (it does not actually erase the win-
     dow's screen image).  Subwindows must  be  deleted  before
     the main window can be deleted.
}
\seealso{
Other window: \code{\link{box}}, \code{\link{mvwin}},
  \code{\link{newwin}}, \code{\link{stdscr}},
  \code{\link{subwin}}
}

