% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{derwin}
\alias{derwin}
\title{create a new sub window with positions relative to its parent window}
\usage{
derwin(win, nlines, ncols, y, x)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{nlines}{the number of lines of the window}

\item{ncols}{the number of columns of the window}

\item{y}{the upper y position}

\item{x}{the left x position}
}
\value{
a new window as an external pointer
}
\description{
Calling derwin is the same as calling subwin, except  that
begin_y and begin_x are relative to the origin of the win-
 dow orig rather than the screen.
}
\seealso{
Other window: \code{\link{box}}, \code{\link{delwin}},
  \code{\link{dupwin}}, \code{\link{mvderwin}},
  \code{\link{mvwin}}, \code{\link{newwin}},
  \code{\link{stdscr}}, \code{\link{subwin}},
  \code{\link{syncok}}, \code{\link{wsyncup}}
}

