% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{dupwin}
\alias{dupwin}
\title{creates  an exact duplicate of a window}
\usage{
dupwin(win)
}
\arguments{
\item{win}{a window object as an external pointer}
}
\value{
a new window as an external pointer
}
\description{
creates  an exact duplicate of a window
}
\seealso{
Other window: \code{\link{delwin}}, \code{\link{derwin}},
  \code{\link{mvderwin}}, \code{\link{mvwin}},
  \code{\link{newwin}}, \code{\link{subwin}},
  \code{\link{syncok}}, \code{\link{wsyncup}}
}

