% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{echo}
\alias{echo}
\title{control whether characters typed by the user are echoed  by  getch()
 as  they  are typed.}
\usage{
echo(bf = TRUE)
}
\arguments{
\item{bf}{whether to enable or disable}
}
\description{
Echoing by the tty driver is always disabled, but
initially getch is in echo mode, so characters  typed  are
echoed.  Authors of most interactive programs prefer to do
their own echoing in a controlled area of the  screen,  or
not  to  echo  at  all, so they disable echoing by calling
noecho.
}
\seealso{
noecho

Other inopts: \code{\link{cbreak}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{keypad}}, \code{\link{meta}},
  \code{\link{nocbreak}}, \code{\link{nodelay}},
  \code{\link{noecho}}, \code{\link{noqiflush}},
  \code{\link{noraw}}, \code{\link{qiflush}},
  \code{\link{raw}}, \code{\link{timeout}}
}

