% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addch.r
\name{echochar}
\alias{echochar}
\title{echochar}
\usage{
echochar(ch, win = stdscr())
}
\arguments{
\item{ch}{a character to print}

\item{win}{a window object as an external pointer}
}
\description{
echochar
}
\seealso{
Other addch: \code{\link{addch}}, \code{\link{mvaddch}},
  \code{\link{mvwaddch}}, \code{\link{waddch}},
  \code{\link{wechochar}}
}

