% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{endwin}
\alias{endwin}
\title{end the curses mode}
\usage{
endwin()
}
\description{
endwin() frees the memory taken by curses sub-system and its data structures
	and puts the terminal in normal mode.
This function must be called after you are done with the curses mode.
}
\seealso{
initscr

Other init: \code{\link{initscr}},
  \code{\link{with_initscr}}
}

