% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.r
\name{ESCDELAY}
\alias{ESCDELAY}
\title{number of ms to wait after reading an escape character}
\usage{
ESCDELAY()
}
\description{
to distinguish between an individual escape character entered on the  keyboard
from escape  sequences  sent  by cursor- and function-keys
}
\seealso{
Other variables: \code{\link{COLORS}},
  \code{\link{COLOR_PAIRS}}, \code{\link{COLS}},
  \code{\link{LINES}}, \code{\link{TABSIZE}},
  \code{\link{curscr}}, \code{\link{newscr}},
  \code{\link{stdscr}}
}

