% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_attron}
\alias{example_attron}
\title{attron example}
\usage{
example_attron(sleep = 5)
}
\arguments{
\item{sleep}{sleep time before exiting}
}
\description{
attron example
}
\seealso{
Other examples: \code{\link{example_attributes}},
  \code{\link{example_box}}, \code{\link{example_getnstr}},
  \code{\link{example_hello}},
  \code{\link{example_mvprintw}},
  \code{\link{example_window}}
}

