% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_color_set}
\alias{example_color_set}
\title{color_set}
\usage{
example_color_set()
}
\description{
color_set
}
\seealso{
Other examples: \code{\link{example_attributes}},
  \code{\link{example_attron}}, \code{\link{example_box}},
  \code{\link{example_change_color}},
  \code{\link{example_color}}, \code{\link{example_getch}},
  \code{\link{example_getnstr}},
  \code{\link{example_hello}},
  \code{\link{example_mouse_debug}},
  \code{\link{example_mouse}},
  \code{\link{example_mvprintw}},
  \code{\link{example_window_border}},
  \code{\link{example_window}}
}

