% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_getnstr}
\alias{example_getnstr}
\title{getnstr example}
\usage{
example_getnstr(sleep = 5, msg = "Just a string")
}
\arguments{
\item{sleep}{sleep time before exiting}

\item{msg}{the message to print in the center of the window}
}
\description{
cf example 4 from NCURSES-Programming-HOWTO
}
\seealso{
Other examples: \code{\link{example_hello}},
  \code{\link{example_mvprintw}}
}

