% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_hello}
\alias{example_hello}
\title{hello world example}
\usage{
example_hello(sleep = 5)
}
\arguments{
\item{sleep}{sleep time before exiting}
}
\description{
hello world example
}
\seealso{
Other examples: \code{\link{example_mvprintw}}
}

