% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_mouse_debug}
\alias{example_mouse_debug}
\title{mouse debug}
\usage{
example_mouse_debug()
}
\description{
mouse debug
}
\seealso{
Other examples: \code{\link{example_attributes}},
  \code{\link{example_attron}}, \code{\link{example_box}},
  \code{\link{example_change_color}},
  \code{\link{example_color_set}},
  \code{\link{example_color}}, \code{\link{example_getch}},
  \code{\link{example_getnstr}},
  \code{\link{example_hello}}, \code{\link{example_mouse}},
  \code{\link{example_mvprintw}},
  \code{\link{example_window_border}},
  \code{\link{example_window}}
}

