% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_mvprintw}
\alias{example_mvprintw}
\title{mvprintw example}
\usage{
example_mvprintw(sleep = 5, msg = "Just a string")
}
\arguments{
\item{sleep}{sleep time before exiting}
}
\description{
mvprintw example
}
\seealso{
Other examples: \code{\link{example_hello}}
}

