% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_window}
\alias{example_window}
\title{window example}
\usage{
example_window(sleep = 5)
}
\arguments{
\item{sleep}{sleep time before exiting}
}
\description{
inspired by http://tldp.org/HOWTO/NCURSES-Programming-HOWTO/windows.html
}
\seealso{
Other examples: \code{\link{example_attributes}},
  \code{\link{example_attron}}, \code{\link{example_box}},
  \code{\link{example_color}}, \code{\link{example_getch}},
  \code{\link{example_getnstr}},
  \code{\link{example_hello}},
  \code{\link{example_mvprintw}},
  \code{\link{example_window_border}}
}

