% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_window_border}
\alias{example_window_border}
\title{window border example}
\usage{
example_window_border()
}
\description{
inspired by http://tldp.org/HOWTO/NCURSES-Programming-HOWTO/windows.html
}
\seealso{
Other examples: \code{\link{example_attributes}},
  \code{\link{example_attron}}, \code{\link{example_box}},
  \code{\link{example_getch}},
  \code{\link{example_getnstr}},
  \code{\link{example_hello}},
  \code{\link{example_mvprintw}},
  \code{\link{example_window}}
}

