% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getch.r
\name{getch}
\alias{getch}
\title{wait for user input}
\usage{
getch()
}
\value{
the keycode, or <= 0 on error or timeout
}
\description{
wait for user input
}
\seealso{
\code{\link{KEYS}}

Other getch: \code{\link{has_key}}, \code{\link{mvgetch}},
  \code{\link{mvwgetch}}, \code{\link{ungetch}},
  \code{\link{wgetch}}
}

