% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{getmaxyx}
\alias{getmaxyx}
\title{get the current beginning coordinates and size of the default window.}
\usage{
getmaxyx(win = stdscr())
}
\arguments{
\item{win}{a window}
}
\value{
a named list of y, x (or #rows, #cols)
}
\description{
get the current beginning coordinates and size of the default window.
}

