% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.r
\name{getmouse_event}
\alias{getmouse_event}
\title{read the event data and pop the event off the queue}
\usage{
getmouse_event()
}
\value{
a R equivalent of the C MEVENT struct as a list, or NULL if no event
}
\description{
read the event data and pop the event off the queue
}
\seealso{
Other mouse: \code{\link{has_mouse}},
  \code{\link{mouse_events}}, \code{\link{mousemask}}
}

