% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input.r
\name{getnstr}
\alias{getnstr}
\title{read character string from terminal from default window}
\usage{
getnstr(nchar_max = 5000L)
}
\arguments{
\item{nchar_max}{max length of the string to read}
}
\value{
the read string
}
\description{
read character string from terminal from default window
}
\seealso{
Other input: \code{\link{getch}}
}

