% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{halfdelay}
\alias{halfdelay}
\title{similar to cbreak mode, but blocks for tenths tenths of seconds}
\usage{
halfdelay(tenths)
}
\arguments{
\item{tenths}{the number of tenths of seconds to block}
}
\value{
FALSE iff nothing  has  been typed during the blocking
}
\description{
Use nocbreak to leave half-delay mode.
}
\seealso{
nocbreak

Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{intrflush}}, \code{\link{keypad}},
  \code{\link{meta}}, \code{\link{nocbreak}},
  \code{\link{nodelay}}, \code{\link{noecho}},
  \code{\link{noqiflush}}, \code{\link{noraw}},
  \code{\link{qiflush}}, \code{\link{raw}},
  \code{\link{timeout}}
}

