% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{has_colors}
\alias{has_colors}
\title{can the terminal manipulate colors}
\usage{
has_colors()
}
\value{
TRUE iff only it can
}
\description{
can the terminal manipulate colors
}
\seealso{
Other color: \code{\link{COLORS}},
  \code{\link{COLOR_PAIR}}, \code{\link{PAIR_NUMBER}},
  \code{\link{assume_default_colors}},
  \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{init_pair}},
  \code{\link{pair_content}}, \code{\link{start_color}},
  \code{\link{use_default_colors}}
}

