% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getch.r
\name{has_key}
\alias{has_key}
\title{tells if the terminal recognizes a key}
\usage{
has_key(ch)
}
\arguments{
\item{ch}{the key code}
}
\value{
TRUE iff the terminal recognizes ch
}
\description{
tells if the terminal recognizes a key
}
\seealso{
\code{\link{KEYS}}

Other getch: \code{\link{getch}}, \code{\link{mvgetch}},
  \code{\link{mvwgetch}}, \code{\link{ungetch}},
  \code{\link{wgetch}}
}

