% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.r
\name{has_mouse}
\alias{has_mouse}
\title{test if the mouse driver has been successfully initialized.}
\usage{
has_mouse()
}
\value{
TRUE iff mouse support is initialized
}
\description{
test if the mouse driver has been successfully initialized.
}
\seealso{
Other mouse: \code{\link{mouse_events}}
}

