% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.r
\name{hline}
\alias{hline}
\title{hline}
\usage{
hline(ch, n, win = stdscr())
}
\arguments{
\item{ch}{a character to draw the line with}

\item{n}{the max nb of characters to draw}

\item{win}{a window object as an external pointer}
}
\description{
hline
}
\seealso{
Other border: \code{\link{border}}, \code{\link{box}},
  \code{\link{vline}}, \code{\link{whline}},
  \code{\link{wvline}}
}

