% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{init_pair}
\alias{init_pair}
\title{changes the definition of a color pair.}
\usage{
init_pair(pair, b, f)
}
\arguments{
\item{pair}{the pair number}

\item{b}{the background color number}

\item{f}{the foreground color number}
}
\description{
changes the definition of a color pair.
}
\seealso{
Other color: \code{\link{COLORS}},
  \code{\link{COLOR_PAIR}}, \code{\link{PAIR_NUMBER}},
  \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{pair_content}}, \code{\link{start_color}}
}

