% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{initscr}
\alias{initscr}
\title{initializes the terminal in curses mode.}
\usage{
initscr()
}
\description{
In some implementations, it clears the screen and presents a blank screen.
}
\seealso{
endwin

Other init: \code{\link{endwin}}, \code{\link{raw}}
}

