% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{intrflush}
\alias{intrflush}
\title{enable/disable the interrupt flush mode}
\usage{
intrflush(bf = TRUE)
}
\arguments{
\item{bf}{whether to enable or disable}
}
\description{
If  the  intrflush  option is enabled (bf is TRUE), and an
       interrupt key  is  pressed  on  the  keyboard  (interrupt,
       break,  quit),  all output in the tty driver queue will be
       flushed, giving the effect of faster response to  the  in-
       terrupt, but causing curses to have the wrong idea of what
       is on the screen.  Disabling the option (bf is FALSE) pre-
       vents  the flush.  The default for the option is inherited
       from the tty driver settings.
}
\seealso{
Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{keypad}},
  \code{\link{meta}}, \code{\link{nocbreak}},
  \code{\link{noecho}}, \code{\link{noqiflush}},
  \code{\link{noraw}}, \code{\link{qiflush}},
  \code{\link{raw}}, \code{\link{timeout}}
}

