% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keys.r
\name{keyname}
\alias{keyname}
\title{return the string corresponding to the key code}
\usage{
keyname(code)
}
\arguments{
\item{code}{an integer code describing a key}
}
\value{
the curses keyname for that key code
}
\description{
N.B: requires \code{\link{keypad}}
}

