% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{keypad}
\alias{keypad}
\title{enable/disable the keypad of the user's terminal.}
\usage{
keypad(win = stdscr(), bf = TRUE)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{bf}{whether to enable or disable}
}
\description{
If enabled (bf is TRUE), the user can press a  func-
       tion  key  (such as an arrow key) and wgetch(3x) returns a
       single  value  representing  the  function  key,   as   in
       KEY_LEFT.   If  disabled  (bf  is  FALSE), curses does not
       treat function keys specially and the program has  to  in-
       terpret the escape sequences itself.  If the keypad in the
       terminal can be turned on (made to transmit) and off (made
       to work locally), turning on this option causes the termi-
       nal keypad to be turned on when wgetch(3x) is called.  The
       default value for keypad is FALSE.
}
\seealso{
Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{meta}}, \code{\link{nocbreak}},
  \code{\link{nodelay}}, \code{\link{noecho}},
  \code{\link{noqiflush}}, \code{\link{noraw}},
  \code{\link{qiflush}}, \code{\link{raw}},
  \code{\link{timeout}}
}

