% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{LINES}
\alias{LINES}
\title{the  height of the screen in characters.}
\usage{
LINES()
}
\value{
the number of lines
}
\description{
the  height of the screen in characters.
}
\seealso{
Other coords: \code{\link{COLS}}, \code{\link{getmaxyx2}},
  \code{\link{getmaxyx}}
}

