% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{meta}
\alias{meta}
\title{enable/disable the terminal to returns 8 (resp.  7) significant
 bits on input}
\usage{
meta(bf = TRUE)
}
\arguments{
\item{bf}{whether to enable or disable}
}
\description{
Initially, whether the terminal returns 7 or 8 significantss a  func-
       bits on input depends on the control mode of the tty driv-and wgetch(3x) returns a
       er  [see  termio(7)].  To force 8 bits to be returned, in-function  key,   as   in
       voke meta(win, TRUE); this is equivalent, under POSIX,  to FALSE), curses does not
       setting  the CS8 flag on the terminal.  To force 7 bits tothe program has  to  in-
       be returned, invoke meta(win, FALSE); this is  equivalent,f.  If the keypad in the
       under POSIX, to setting the CS7 flag on the terminal.  The transmit) and off (made
       window argument, win, is always ignored.  If the  terminfooption causes the termi-
       capabilities  smm (meta_on) and rmm (meta_off) are definedetch(3x) is called.  The
       for the  terminal,  smm  is  sent  to  the  terminal  when
       meta(win,  TRUE)  is called and rmm is sent when meta(win,
       FALSE) is called.
}
\seealso{
Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{keypad}}, \code{\link{nocbreak}},
  \code{\link{nodelay}}, \code{\link{noecho}},
  \code{\link{noqiflush}}, \code{\link{noraw}},
  \code{\link{qiflush}}, \code{\link{raw}},
  \code{\link{timeout}}
}

