% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.r
\docType{data}
\name{MOUSE_EVENT_NAMES}
\alias{MOUSE_EVENT_NAMES}
\title{BUTTON5_PRESSED',
BUTTON5_RELEASED',
BUTTON5_CLICKED',
BUTTON5_DOUBLE_CLICKED',
BUTTON5_TRIPLE_CLICKED',}
\format{An object of class \code{character} of length 25.}
\usage{
MOUSE_EVENT_NAMES
}
\description{
BUTTON5_PRESSED',
BUTTON5_RELEASED',
BUTTON5_CLICKED',
BUTTON5_DOUBLE_CLICKED',
BUTTON5_TRIPLE_CLICKED',
}
\keyword{datasets}

