% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mouse.r
\name{mousemask}
\alias{mousemask}
\title{set the mouse events to  be  reported.}
\usage{
mousemask(mask = bitwOr(MOUSE$ALL_MOUSE_EVENTS, MOUSE$REPORT_MOUSE_POSITION))
}
\value{
a named list of effective, old mask:
  a mask to indicate which  of  the  specified  mouse events can be reported,
 and the previous mask
}
\description{
To  make mouse events visible, use the mousemask function.
This will set the mouse events to  be  reported.   By  de-
       fault,  no  mouse  events are reported.  The function will
       return a mask to indicate which  of  the  specified  mouse
       events  can be reported; on complete failure it returns 0.
       If oldmask is non-NULL, this function fills the  indicated
       location  with  the  previous  value of the given window's
      mouse event mask.
}
\details{
As a side effect, setting a zero mousemask  may  turn  off
       the  mouse pointer; setting a nonzero mask may turn it on.
       Whether this happens is device-dependent.
}
\seealso{
Other mouse: \code{\link{getmouse_event}},
  \code{\link{has_mouse}}, \code{\link{mouse_events}}
}

