% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addch.r
\name{mvaddch}
\alias{mvaddch}
\title{mvaddch}
\usage{
mvaddch(y, x, ch, win = stdscr())
}
\arguments{
\item{y}{the y position (aka the row)}

\item{x}{the x position (aka the column)}

\item{ch}{a character to print}

\item{win}{a window object as an external pointer}
}
\description{
mvaddch
}
\seealso{
Other output: \code{\link{addch}}, \code{\link{clear}},
  \code{\link{mvprintw}}, \code{\link{mvwaddch}},
  \code{\link{printw}}, \code{\link{refresh}},
  \code{\link{waddch}}, \code{\link{wrefresh}}
}

