% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{mvderwin}
\alias{mvderwin}
\title{moves a derived window (or subwindow) inside its parent window.}
\usage{
mvderwin(win, y, x)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{y}{the upper y position}

\item{x}{the left x position}
}
\value{
TRUE iff the window has been moved
}
\description{
The screen-relative parameters of the  window are not changed.
This routine is used to display different parts of the  parent  window
 at  the  same physical position on the screen.
}
\seealso{
Other window: \code{\link{box}}, \code{\link{delwin}},
  \code{\link{derwin}}, \code{\link{dupwin}},
  \code{\link{mvwin}}, \code{\link{newwin}},
  \code{\link{stdscr}}, \code{\link{subwin}}
}

