% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getch.r
\name{mvgetch}
\alias{mvgetch}
\title{move and wait for user input}
\usage{
mvgetch(y, x)
}
\arguments{
\item{y}{the y position (aka the row)}

\item{x}{the x position (aka the column)}
}
\value{
the keycode, or <= 0 on error or timeout
}
\description{
move and wait for user input
}
\seealso{
\code{\link{KEYS}}

Other getch: \code{\link{getch}}, \code{\link{has_key}},
  \code{\link{mvwgetch}}, \code{\link{ungetch}},
  \code{\link{wgetch}}
}

