% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{mvprintw}
\alias{mvprintw}
\title{perform a cursor movement using wmove then printw()}
\usage{
mvprintw(y, x, str, ...)
}
\arguments{
\item{y}{the y position (aka the row)}

\item{x}{the x position (aka the column)}

\item{str}{a string to print via sprintf}

\item{...}{additional args for sprintf}
}
\description{
may die if the position is outside the window
}
\seealso{
Other output: \code{\link{addch}}, \code{\link{clear}},
  \code{\link{mvaddch}}, \code{\link{mvwaddch}},
  \code{\link{printw}}, \code{\link{refresh}},
  \code{\link{waddch}}, \code{\link{wrefresh}}
}

