% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addch.r
\name{mvwaddch}
\alias{mvwaddch}
\title{mvwaddch}
\usage{
mvwaddch(win, y, x, ch)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{y}{the y position (aka the row)}

\item{x}{the x position (aka the column)}

\item{ch}{a character to print}
}
\description{
mvwaddch
}
\seealso{
Other addch: \code{\link{addch}}, \code{\link{echochar}},
  \code{\link{mvaddch}}, \code{\link{waddch}},
  \code{\link{wechochar}}
}

