% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{mvwin}
\alias{mvwin}
\title{moves the window so that the upper left-hand corner is at position (x, y)}
\usage{
mvwin(win, y, x)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{y}{the upper y position}

\item{x}{the left x position}
}
\value{
TRUE iff the window has been moved
}
\description{
If the move would cause the
      window to be off the screen, it is an error and the window
      is not moved.  Moving subwindows is allowed, but should be
      avoided.
}
\seealso{
Other window: \code{\link{border}}, \code{\link{box}},
  \code{\link{delwin}}, \code{\link{derwin}},
  \code{\link{dupwin}}, \code{\link{mvderwin}},
  \code{\link{newwin}}, \code{\link{stdscr}},
  \code{\link{subwin}}, \code{\link{syncok}},
  \code{\link{wsyncup}}
}

