% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{ncurses_attributes}
\alias{ncurses_attributes}
\title{get the attributes values}
\usage{
ncurses_attributes()
}
\description{
get the attributes values
}
\seealso{
Other attributes: \code{\link{COLOR_PAIR}},
  \code{\link{attroff}}, \code{\link{attron}},
  \code{\link{with_attr}}
}

