% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.r
\name{newscr}
\alias{newscr}
\title{special window newscr to  hold  updates  to  the terminal screen}
\usage{
newscr()
}
\value{
newscr as an external pointer
}
\description{
special window newscr to  hold  updates  to  the terminal screen
}
\seealso{
Other variables: \code{\link{COLORS}},
  \code{\link{COLOR_PAIRS}}, \code{\link{COLS}},
  \code{\link{ESCDELAY}}, \code{\link{LINES}},
  \code{\link{TABSIZE}}, \code{\link{curscr}},
  \code{\link{stdscr}}
}

