% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{nocbreak}
\alias{nocbreak}
\title{returns the terminal to normal (cooked) mode.}
\usage{
nocbreak()
}
\description{
returns the terminal to normal (cooked) mode.
}
\seealso{
cbreak

Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{keypad}}, \code{\link{meta}},
  \code{\link{noecho}}, \code{\link{noqiflush}},
  \code{\link{noraw}}, \code{\link{qiflush}},
  \code{\link{raw}}, \code{\link{timeout}}
}

