% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{nodelay}
\alias{nodelay}
\title{cause getch to be a non-blocking call}
\usage{
nodelay(bf = TRUE, win = stdscr())
}
\arguments{
\item{bf}{whether to enable or disable}

\item{win}{a window object as an external pointer}
}
\value{
FALSE iff nothing  has  been typed during the blocking
}
\description{
If  no input is ready, getch returns ERR.  If disabled (bf
    is FALSE), getch waits until a key is pressed.
}
\details{
While interpreting an input  escape  sequence,  wgetch(3x)
sets a timer while waiting for the next character.  If no-
  timeout(win, TRUE) is called, then wgetch does not  set  a
timer.  The purpose of the timeout is to differentiate be-
  tween sequences received from a  function  key  and  those
typed by a user.
}
\seealso{
getch

Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{keypad}}, \code{\link{meta}},
  \code{\link{nocbreak}}, \code{\link{noecho}},
  \code{\link{noqiflush}}, \code{\link{noraw}},
  \code{\link{qiflush}}, \code{\link{raw}},
  \code{\link{timeout}}
}

