% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{noecho}
\alias{noecho}
\title{place the terminal out of raw mode.
#'}
\usage{
noecho()
}
\description{
place the terminal out of raw mode.
#'
}
\seealso{
echo

Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{keypad}}, \code{\link{meta}},
  \code{\link{nocbreak}}, \code{\link{noqiflush}},
  \code{\link{noraw}}, \code{\link{qiflush}},
  \code{\link{raw}}, \code{\link{timeout}}
}

