% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{noqiflush}
\alias{noqiflush}
\title{normal flush of  input and  output queues associated with the INTR, QUIT and SUSP
   characters will not be done}
\usage{
noqiflush()
}
\description{
normal flush of  input and  output queues associated with the INTR, QUIT and SUSP
   characters will not be done
}
\seealso{
qiflush

Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{keypad}}, \code{\link{meta}},
  \code{\link{nocbreak}}, \code{\link{nodelay}},
  \code{\link{noecho}}, \code{\link{noraw}},
  \code{\link{qiflush}}, \code{\link{raw}},
  \code{\link{timeout}}
}

