% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{noraw}
\alias{noraw}
\title{place the terminal out of raw mode.
#'}
\usage{
noraw()
}
\description{
place the terminal out of raw mode.
#'
}
\seealso{
raw

Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{keypad}}, \code{\link{meta}},
  \code{\link{nocbreak}}, \code{\link{noecho}},
  \code{\link{noqiflush}}, \code{\link{qiflush}},
  \code{\link{raw}}, \code{\link{timeout}}
}

