% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{pair_content}
\alias{pair_content}
\title{find out what colors  a given color-pair consists of}
\usage{
pair_content(pair)
}
\arguments{
\item{pair}{the pair number}
}
\value{
the colors (foreground, background) as a named list (F, B)
}
\description{
find out what colors  a given color-pair consists of
}
\seealso{
Other color: \code{\link{COLORS}},
  \code{\link{COLOR_PAIR}}, \code{\link{PAIR_NUMBER}},
  \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{start_color}}
}

