% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{PAIR_NUMBER}
\alias{PAIR_NUMBER}
\title{extracts the color value for a video attribute}
\usage{
PAIR_NUMBER(vattr)
}
\arguments{
\item{vattr}{a video attribute}
}
\value{
the pair number
}
\description{
inverse of \code{\link{COLOR_PAIR}}
}
\seealso{
Other color: \code{\link{COLORS}},
  \code{\link{COLOR_PAIR}}, \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{pair_content}},
  \code{\link{start_color}}
}

