% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.r
\name{printw}
\alias{printw}
\title{work much like printf()}
\usage{
printw(str)
}
\arguments{
\item{str}{a string to print}
}
\description{
If you want to move the cursor first and then print using printw() function,
use move() first and then use printw()
}
\seealso{
Other output: \code{\link{addch}}, \code{\link{clear}},
  \code{\link{mvaddch}}, \code{\link{mvprintw}},
  \code{\link{mvwaddch}}, \code{\link{refresh}},
  \code{\link{waddch}}, \code{\link{wrefresh}}
}

