% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{qiflush}
\alias{qiflush}
\title{output queues associated with the INTR, QUIT and SUSP will be flushed when these control
characters are read}
\usage{
qiflush(bf = TRUE)
}
\arguments{
\item{bf}{whether to enable or disable}
}
\description{
You may want to call noqiflush in  a
       signal  handler  if  you want output to continue as though
       the interrupt had not occurred, after the handler exits.
}
\seealso{
noqiflush

Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{keypad}}, \code{\link{meta}},
  \code{\link{nocbreak}}, \code{\link{nodelay}},
  \code{\link{noecho}}, \code{\link{noqiflush}},
  \code{\link{noraw}}, \code{\link{raw}},
  \code{\link{timeout}}
}

