% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{raw}
\alias{raw}
\title{disable line buffering and signals by keystrokes}
\usage{
raw()
}
\description{
Normally the terminal driver buffers the characters a user types
until a new line or carriage return is encountered.
This function disables this buffering.
}
\details{
Control characters like suspend (CTRL-Z), interrupt and quit (CTRL-C)
are directly passed to the program without generating a signal.
}
\seealso{
cbreak

Other init: \code{\link{endwin}}, \code{\link{initscr}}
}

