% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\docType{data}
\name{rcurses.A_CHARTEXT}
\alias{rcurses.A_CHARTEXT}
\alias{rcurses.A_CHARTEXT}
\title{Interface to rcurses.A_CHARTEXT}
\format{An object of class \code{integer} of length 1.}
\usage{
rcurses.A_CHARTEXT

rcurses.A_CHARTEXT
}
\value{
value from corresponding ncurses C constant

value from corresponding ncurses C constant
}
\description{
Interface to rcurses.A_CHARTEXT

Interface to rcurses.A_CHARTEXT
}
\keyword{datasets}
