% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.addch}
\alias{rcurses.addch}
\title{Interface to rcurses.addch}
\usage{
rcurses.addch(window, ch, y = NULL, x = NULL, attr = NULL)
}
\arguments{
\item{window}{window returned by call to initscr}

\item{ch}{newly painted character}

\item{y}{y position of newly painted character}

\item{x}{x position of newly painted character}

\item{attr}{attribute of newly painted character}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.addch
}
