% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.addstr}
\alias{rcurses.addstr}
\title{Interface to rcurses.addnstr}
\usage{
rcurses.addstr(window, str, y = NULL, x = NULL, attr = NULL)
}
\arguments{
\item{window}{window returned by call to initscr}

\item{str}{newly painted string}

\item{y}{starting y position of newly painted string}

\item{x}{starting y position of newly painted string}

\item{attr}{attribute of newly painted character}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.addnstr
}
