% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.baudrate}
\alias{rcurses.baudrate}
\title{Interface to rcurses.baudrate}
\usage{
rcurses.baudrate()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.baudrate
}
