% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.bkgd}
\alias{rcurses.bkgd}
\title{Interface to rcurses.bkgd}
\usage{
rcurses.bkgd(window, ch, attr = NULL)
}
\arguments{
\item{window}{window returned by call to initscr}

\item{ch}{new background property of the window}

\item{attr}{attributes of background property}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.bkgd
}
