% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.can_change_color}
\alias{rcurses.can_change_color}
\title{Interface to rcurses.can_change_color}
\usage{
rcurses.can_change_color()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.can_change_color
}
