% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.chgat}
\alias{rcurses.chgat}
\title{Interface to rcurses.chgat}
\usage{
rcurses.chgat(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.chgat
}
