% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.clear}
\alias{rcurses.clear}
\title{Interface to rcurses.clear}
\usage{
rcurses.clear(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.clear
}
