% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.clearok}
\alias{rcurses.clearok}
\title{Interface to rcurses.clearok}
\usage{
rcurses.clearok(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.clearok
}
