% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.color_content}
\alias{rcurses.color_content}
\title{Interface to rcurses.color_content}
\usage{
rcurses.color_content()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.color_content
}
