% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.color_pair}
\alias{rcurses.color_pair}
\title{Interface to rcurses.color_pair}
\usage{
rcurses.color_pair(color_number)
}
\arguments{
\item{color_number}{color pair ID returned by init_pair}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.color_pair
}
