% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.curs_set}
\alias{rcurses.curs_set}
\title{Interface to rcurses.curs_set}
\usage{
rcurses.curs_set()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.curs_set
}
