% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.cursyncup}
\alias{rcurses.cursyncup}
\title{Interface to rcurses.cursyncup}
\usage{
rcurses.cursyncup(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.cursyncup
}
