% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.def_prog_mode}
\alias{rcurses.def_prog_mode}
\title{Interface to rcurses.def_prog_mode}
\usage{
rcurses.def_prog_mode()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.def_prog_mode
}
