% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.delay_output}
\alias{rcurses.delay_output}
\title{Interface to rcurses.delay_output}
\usage{
rcurses.delay_output()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.delay_output
}
