% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.delch}
\alias{rcurses.delch}
\title{Interface to rcurses.delch}
\usage{
rcurses.delch(window, y = NULL, x = NULL)
}
\arguments{
\item{window}{window returned by call to initscr}

\item{y}{y position of deleted character}

\item{x}{x position of deleted character}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.delch
}
