% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.deleteln}
\alias{rcurses.deleteln}
\title{Interface to rcurses.deleteln}
\usage{
rcurses.deleteln(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.deleteln
}
