% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.doupdate}
\alias{rcurses.doupdate}
\title{Interface to rcurses.doupdate}
\usage{
rcurses.doupdate()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.doupdate
}
