% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.echochar}
\alias{rcurses.echochar}
\title{Interface to rcurses.echochar}
\usage{
rcurses.echochar(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.echochar
}
