% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.enclose}
\alias{rcurses.enclose}
\title{Interface to rcurses.enclose}
\usage{
rcurses.enclose(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.enclose
}
