% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.erasechar}
\alias{rcurses.erasechar}
\title{Interface to rcurses.erasechar}
\usage{
rcurses.erasechar()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.erasechar
}
