% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.flushinp}
\alias{rcurses.flushinp}
\title{Interface to rcurses.flushinp}
\usage{
rcurses.flushinp()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.flushinp
}
