% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.get_wch}
\alias{rcurses.get_wch}
\title{Interface to rcurses.get_wch}
\usage{
rcurses.get_wch(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.get_wch
}
