% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.getbegyx}
\alias{rcurses.getbegyx}
\title{Interface to rcurses.getbegyx}
\usage{
rcurses.getbegyx(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.getbegyx
}
