% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.getch}
\alias{rcurses.getch}
\title{Interface to rcurses.getch}
\usage{
rcurses.getch(window, y = NULL, x = NULL)
}
\arguments{
\item{window}{window returned by call to initscr}

\item{y}{y position of newly painted character}

\item{x}{x position of newly painted character}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.getch
}
