% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.getkey}
\alias{rcurses.getkey}
\title{Interface to rcurses.getkey}
\usage{
rcurses.getkey(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.getkey
}
