% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.getmouse}
\alias{rcurses.getmouse}
\title{Interface to rcurses.getmouse}
\usage{
rcurses.getmouse()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.getmouse
}
