% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.getstr}
\alias{rcurses.getstr}
\title{Interface to rcurses.getstr}
\usage{
rcurses.getstr(window, n = NULL, y = NULL, x = NULL)
}
\arguments{
\item{window}{window returned by call to initscr}

\item{n}{maximum number of characters to paint}

\item{y}{starting y position of newly painted string}

\item{x}{starting y position of newly painted string}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.getstr
}
