% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.getyx}
\alias{rcurses.getyx}
\title{Interface to rcurses.getyx}
\usage{
rcurses.getyx(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.getyx
}
