% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.halfdelay}
\alias{rcurses.halfdelay}
\title{Interface to rcurses.halfdelay}
\usage{
rcurses.halfdelay()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.halfdelay
}
