% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.has_colors}
\alias{rcurses.has_colors}
\title{Interface to rcurses.has_colors}
\usage{
rcurses.has_colors()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.has_colors
}
