% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.hline}
\alias{rcurses.hline}
\title{Interface to rcurses.hline}
\usage{
rcurses.hline(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.hline
}
