% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.idcok}
\alias{rcurses.idcok}
\title{Interface to rcurses.idcok}
\usage{
rcurses.idcok(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.idcok
}
