% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.idlok}
\alias{rcurses.idlok}
\title{Interface to rcurses.idlok}
\usage{
rcurses.idlok(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.idlok
}
