% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.immedok}
\alias{rcurses.immedok}
\title{Interface to rcurses.immedok}
\usage{
rcurses.immedok(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.immedok
}
