% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.inch}
\alias{rcurses.inch}
\title{Interface to rcurses.inch}
\usage{
rcurses.inch(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.inch
}
