% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.init_color}
\alias{rcurses.init_color}
\title{Interface to rcurses.init_color}
\usage{
rcurses.init_color()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.init_color
}
