% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.init_pair}
\alias{rcurses.init_pair}
\title{Interface to rcurses.init_pair}
\usage{
rcurses.init_pair(pair_number, fg, bg)
}
\arguments{
\item{pair_number}{ID representing this new color pair}

\item{fg}{foreground color (either color macro or created with init_color)}

\item{bg}{background color (either color macro or created with init_color)}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.init_pair
}
