% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.insch}
\alias{rcurses.insch}
\title{Interface to rcurses.insch}
\usage{
rcurses.insch(window, ch, y = NULL, x = NULL, attr = NULL)
}
\arguments{
\item{window}{window returned by call to initscr}

\item{ch}{newly inserted character}

\item{y}{y position of newly inserted character}

\item{x}{x position of newly inserted character}

\item{attr}{attribute of newly painted character}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.insch
}
