% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.insdelln}
\alias{rcurses.insdelln}
\title{Interface to rcurses.insdelln}
\usage{
rcurses.insdelln(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.insdelln
}
