% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.is_linetouched}
\alias{rcurses.is_linetouched}
\title{Interface to rcurses.is_linetouched}
\usage{
rcurses.is_linetouched(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.is_linetouched
}
