% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.is_term_resized}
\alias{rcurses.is_term_resized}
\title{Interface to rcurses.is_term_resized}
\usage{
rcurses.is_term_resized()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.is_term_resized
}
