% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.isendwin}
\alias{rcurses.isendwin}
\title{Interface to rcurses.isendwin}
\usage{
rcurses.isendwin()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.isendwin
}
