% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\docType{data}
\name{rcurses.KEY_NEXT}
\alias{rcurses.KEY_NEXT}
\title{Interface to rcurses.KEY_NEXT}
\format{An object of class \code{integer} of length 1.}
\usage{
rcurses.KEY_NEXT
}
\value{
value from corresponding ncurses C constant
}
\description{
Interface to rcurses.KEY_NEXT
}
\keyword{datasets}
