% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.keypad}
\alias{rcurses.keypad}
\title{Interface to rcurses.keypad}
\usage{
rcurses.keypad(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.keypad
}
