% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.killchar}
\alias{rcurses.killchar}
\title{Interface to rcurses.killchar}
\usage{
rcurses.killchar()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.killchar
}
