% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.leaveok}
\alias{rcurses.leaveok}
\title{Interface to rcurses.leaveok}
\usage{
rcurses.leaveok(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.leaveok
}
