% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.longname}
\alias{rcurses.longname}
\title{Interface to rcurses.longname}
\usage{
rcurses.longname()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.longname
}
