% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.mouseinterval}
\alias{rcurses.mouseinterval}
\title{Interface to rcurses.mouseinterval}
\usage{
rcurses.mouseinterval()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.mouseinterval
}
