% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.mousemask}
\alias{rcurses.mousemask}
\title{Interface to rcurses.mousemask}
\usage{
rcurses.mousemask()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.mousemask
}
