% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.move}
\alias{rcurses.move}
\title{Interface to rcurses.move}
\usage{
rcurses.move(window, new_y, new_x)
}
\arguments{
\item{window}{window returned by call to initscr}

\item{new_y}{new y position of cursor}

\item{new_x}{new x position of cursor}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.move
}
