% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.mvderwin}
\alias{rcurses.mvderwin}
\title{Interface to rcurses.mvderwin}
\usage{
rcurses.mvderwin(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.mvderwin
}
