% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.mvwin}
\alias{rcurses.mvwin}
\title{Interface to rcurses.mvwin}
\usage{
rcurses.mvwin(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.mvwin
}
