% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.nocbreak}
\alias{rcurses.nocbreak}
\title{Interface to rcurses.nocbreak}
\usage{
rcurses.nocbreak()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.nocbreak
}
