% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.nodelay}
\alias{rcurses.nodelay}
\title{Interface to rcurses.nodelay}
\usage{
rcurses.nodelay(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.nodelay
}
