% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.noqiflush}
\alias{rcurses.noqiflush}
\title{Interface to rcurses.noqiflush}
\usage{
rcurses.noqiflush()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.noqiflush
}
