% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.notimeout}
\alias{rcurses.notimeout}
\title{Interface to rcurses.notimeout}
\usage{
rcurses.notimeout(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.notimeout
}
