% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.noutrefresh}
\alias{rcurses.noutrefresh}
\title{Interface to rcurses.noutrefresh}
\usage{
rcurses.noutrefresh(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.noutrefresh
}
