% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.overlay}
\alias{rcurses.overlay}
\title{Interface to rcurses.overlay}
\usage{
rcurses.overlay(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.overlay
}
