% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.pair_content}
\alias{rcurses.pair_content}
\title{Interface to rcurses.pair_content}
\usage{
rcurses.pair_content()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.pair_content
}
