% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.pair_number}
\alias{rcurses.pair_number}
\title{Interface to rcurses.pair_number}
\usage{
rcurses.pair_number()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.pair_number
}
