% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.putwin}
\alias{rcurses.putwin}
\title{Interface to rcurses.putwin}
\usage{
rcurses.putwin(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.putwin
}
