\name{rcurses}
\alias{rcurses}
\title{R Interface to curses Library}
\description{
R interface to standard curses library, which is used to move the
cursor around the screen, write/erase character strings and so on.
See curses function list at

https://doc.bccnsoft.com/docs/php-docs-7-en/ref.ncurses.html

and tutorial at

http://heather.cs.ucdavis.edu/~matloff/UnixAndC/CLanguage/Curses.pdf

Also, the README.md file in this package gives an introduction to
rcurses.
}
