% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.redrawwin}
\alias{rcurses.redrawwin}
\title{Interface to rcurses.redrawwin}
\usage{
rcurses.redrawwin(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.redrawwin
}
