% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.refresh}
\alias{rcurses.refresh}
\title{Interface to rcurses.refresh}
\usage{
rcurses.refresh(window, pminrow = NULL, pmincol = NULL, sminrow = NULL,
  smincol = NULL, smaxrow = NULL, smaxcol = NULL)
}
\arguments{
\item{window}{window returned by call to initscr}

\item{pminrow}{specify the upper left-hand corner of the rectangle to be displayed in the pad}

\item{pmincol}{specify the upper left-hand corner of the rectangle to be displayed in the pad}

\item{sminrow}{specify the edges of the rectangle to be displayed on the screen}

\item{smincol}{specify the edges of the rectangle to be displayed on the screen}

\item{smaxrow}{specify the edges of the rectangle to be displayed on the screen}

\item{smaxcol}{specify the edges of the rectangle to be displayed on the screen}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.refresh
}
