% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.reset_shell_mode}
\alias{rcurses.reset_shell_mode}
\title{Interface to rcurses.reset_shell_mode}
\usage{
rcurses.reset_shell_mode()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.reset_shell_mode
}
