% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.resize_term}
\alias{rcurses.resize_term}
\title{Interface to rcurses.resize_term}
\usage{
rcurses.resize_term()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.resize_term
}
