% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.resizeterm}
\alias{rcurses.resizeterm}
\title{Interface to rcurses.resizeterm}
\usage{
rcurses.resizeterm()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.resizeterm
}
