% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.setupterm}
\alias{rcurses.setupterm}
\title{Interface to rcurses.setupterm}
\usage{
rcurses.setupterm()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.setupterm
}
