% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.standend}
\alias{rcurses.standend}
\title{Interface to rcurses.standend}
\usage{
rcurses.standend(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.standend
}
