% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.start_color}
\alias{rcurses.start_color}
\title{Interface to rcurses.start_color}
\usage{
rcurses.start_color()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.start_color
}
