% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.syncdown}
\alias{rcurses.syncdown}
\title{Interface to rcurses.syncdown}
\usage{
rcurses.syncdown(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.syncdown
}
