% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.syncok}
\alias{rcurses.syncok}
\title{Interface to rcurses.syncok}
\usage{
rcurses.syncok(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.syncok
}
