% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.syncup}
\alias{rcurses.syncup}
\title{Interface to rcurses.syncup}
\usage{
rcurses.syncup(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.syncup
}
