% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.termattrs}
\alias{rcurses.termattrs}
\title{Interface to rcurses.termattrs}
\usage{
rcurses.termattrs()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.termattrs
}
