% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.tigetflag}
\alias{rcurses.tigetflag}
\title{Interface to rcurses.tigetflag}
\usage{
rcurses.tigetflag()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.tigetflag
}
