% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.tigetnum}
\alias{rcurses.tigetnum}
\title{Interface to rcurses.tigetnum}
\usage{
rcurses.tigetnum()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.tigetnum
}
