% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.tigetstr}
\alias{rcurses.tigetstr}
\title{Interface to rcurses.tigetstr}
\usage{
rcurses.tigetstr()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.tigetstr
}
