% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.timeout}
\alias{rcurses.timeout}
\title{Interface to rcurses.timeout}
\usage{
rcurses.timeout(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.timeout
}
