% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.touchwin}
\alias{rcurses.touchwin}
\title{Interface to rcurses.touchwin}
\usage{
rcurses.touchwin(window)
}
\arguments{
\item{window}{window returned by call to initscr}
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.touchwin
}
