% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.typeahead}
\alias{rcurses.typeahead}
\title{Interface to rcurses.typeahead}
\usage{
rcurses.typeahead()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.typeahead
}
