% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.unget_wch}
\alias{rcurses.unget_wch}
\title{Interface to rcurses.unget_wch}
\usage{
rcurses.unget_wch()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.unget_wch
}
