% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.ungetmouse}
\alias{rcurses.ungetmouse}
\title{Interface to rcurses.ungetmouse}
\usage{
rcurses.ungetmouse()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.ungetmouse
}
