% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.update_lines_cols}
\alias{rcurses.update_lines_cols}
\title{Interface to rcurses.update_lines_cols}
\usage{
rcurses.update_lines_cols()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.update_lines_cols
}
