% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\name{rcurses.use_default_colors}
\alias{rcurses.use_default_colors}
\title{Interface to rcurses.use_default_colors}
\usage{
rcurses.use_default_colors()
}
\value{
value from corresponding ncurses C function
}
\description{
Interface to rcurses.use_default_colors
}
