% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcurses.R
\docType{data}
\name{rcurses.version}
\alias{rcurses.version}
\title{Interface to rcurses.version}
\format{An object of class \code{character} of length 1.}
\usage{
rcurses.version
}
\value{
value from corresponding ncurses C constant
}
\description{
Interface to rcurses.version
}
\keyword{datasets}
