% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.r
\name{refresh}
\alias{refresh}
\title{refresh the default window}
\usage{
refresh()
}
\description{
refresh the default window
}
\seealso{
Other output: \code{\link{clear}}, \code{\link{mvprintw}},
  \code{\link{printw}}, \code{\link{wrefresh}}
}

