% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{start_color}
\alias{start_color}
\title{allocate the color table}
\usage{
start_color()
}
\description{
allocate the color table
}
\seealso{
Other color: \code{\link{COLOR_PAIR}}, \code{\link{COLOR}},
  \code{\link{PAIR_NUMBER}},
  \code{\link{assume_default_colors}},
  \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{pair_content}},
  \code{\link{use_default_colors}}
}

