% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{subwin}
\alias{subwin}
\title{create a new sub window}
\usage{
subwin(win, nlines, ncols, y, x)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{nlines}{the number of lines of the window}

\item{ncols}{the number of columns of the window}

\item{y}{the upper y position}

\item{x}{the left x position}
}
\value{
a new window as an external pointer
}
\description{
with defaults create a full-screen window
The subwindow shares memory with  the  window
      orig,  so that changes made to one window will affect both
      windows.  When using this routine, it is necessary to call
      touchwin  or  touchline on orig before calling wrefresh on
     the subwindow.
}
\seealso{
Other window: \code{\link{delwin}}, \code{\link{derwin}},
  \code{\link{dupwin}}, \code{\link{mvderwin}},
  \code{\link{mvwin}}, \code{\link{newwin}},
  \code{\link{syncok}}, \code{\link{wsyncup}}
}

