% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.r
\name{TABSIZE}
\alias{TABSIZE}
\title{number  of columns for a tab}
\usage{
TABSIZE()
}
\description{
This  variable  holds  the  number  of columns used by the
curses library when converting a tab character  to  spaces
as it adds the tab to a window
}
\seealso{
Other variables: \code{\link{COLORS}},
  \code{\link{COLOR_PAIRS}}, \code{\link{COLS}},
  \code{\link{ESCDELAY}}, \code{\link{LINES}},
  \code{\link{curscr}}, \code{\link{newscr}},
  \code{\link{stdscr}}
}

