% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getch.r
\name{ungetch}
\alias{ungetch}
\title{wait for user input}
\usage{
ungetch(ch)
}
\arguments{
\item{ch}{the key code}
}
\description{
die if there is no more room in the FIFO
}
\seealso{
\code{\link{KEYS}}

Other getch: \code{\link{getch}}, \code{\link{has_key}},
  \code{\link{mvgetch}}, \code{\link{mvwgetch}},
  \code{\link{wgetch}}
}

