% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{use_default_colors}
\alias{use_default_colors}
\title{assign terminal default foreground/background colors to color number  -1.}
\usage{
use_default_colors()
}
\description{
die if the terminal does not support this extension
}
\seealso{
Other color: \code{\link{COLOR_PAIR}}, \code{\link{COLOR}},
  \code{\link{PAIR_NUMBER}},
  \code{\link{assume_default_colors}},
  \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{pair_content}},
  \code{\link{start_color}}
}

