% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\docType{data}
\name{ATTRIBUTE_NAMES}
\alias{ATTRIBUTE_NAMES}
\alias{A_ALTCHARSET}
\alias{A_BLINK}
\alias{A_BOLD}
\alias{A_CHARTEXT}
\alias{A_DIM}
\alias{A_INVIS}
\alias{A_NORMAL}
\alias{A_PROTECT}
\alias{A_REVERSE}
\alias{A_STANDOUT}
\alias{A_UNDERLINE}
\title{VIDEO ATTRIBUTES}
\format{An object of class \code{character} of length 11.}
\usage{
ATTRIBUTE_NAMES
}
\description{
A_NORMAL       Normal display (no highlight)
              A_STANDOUT     Best highlighting mode of the terminal.
             A_UNDERLINE    Underlining
              A_REVERSE      Reverse video
              A_BLINK        Blinking
              A_DIM          Half bright
              A_BOLD         Extra bright or bold
              A_PROTECT      Protected mode
              A_INVIS        Invisible or blank mode
              A_ALTCHARSET   Alternate character set
              A_ITALIC       Italics (non-X/Open extension)
              A_CHARTEXT     Bit-mask to extract a character
}
\keyword{datasets}

