% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.r
\name{vline}
\alias{vline}
\title{vline}
\usage{
vline(ch, n, win = stdscr())
}
\arguments{
\item{ch}{a character to draw the line with}

\item{n}{the max nb of characters to draw}

\item{win}{a window object as an external pointer}
}
\description{
vline
}
\seealso{
Other border: \code{\link{border}}, \code{\link{box}},
  \code{\link{wvline}}
}

