% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addch.r
\name{wechochar}
\alias{wechochar}
\title{optimized waddch + refresh}
\usage{
wechochar(win, ch)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{ch}{a character to print}
}
\description{
The  echochar  and  wechochar routines are equivalent to a
call to addch followed by a call to refresh(3x), or a call
to  waddch  followed by a call to wrefresh.  The knowledge
that only a single character is being output is used  and,
for  non-control  characters,  a  considerable performance
gain may be seen by using these routines instead of  their
equivalents.
}
\seealso{
Other addch: \code{\link{addch}}, \code{\link{echochar}},
  \code{\link{mvaddch}}, \code{\link{mvwaddch}},
  \code{\link{waddch}}
}

