% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getch.r
\name{wgetch}
\alias{wgetch}
\title{wait for user input}
\usage{
wgetch(win)
}
\arguments{
\item{win}{a window object as an external pointer}
}
\value{
the keycode, or <= 0 on error or timeout
}
\description{
wait for user input
}
\seealso{
\code{\link{KEYS}}

Other getch: \code{\link{getch}}, \code{\link{has_key}}
}

