% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.r
\name{whline}
\alias{whline}
\title{draw  an horizontal  (top  to bottom) line}
\usage{
whline(win, ch, n)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{ch}{a character to draw the line with}

\item{n}{the max nb of characters to draw}
}
\description{
draw  an horizontal  (top  to bottom) line
}
\seealso{
wvline

Other border: \code{\link{border}}, \code{\link{box}},
  \code{\link{hline}}, \code{\link{vline}},
  \code{\link{wvline}}
}

