% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{window_params}
\alias{window_params}
\title{window params}
\arguments{
\item{win}{a window object as an external pointer}

\item{nlines}{the number of lines of the window}

\item{ncols}{the number of columns of the window}

\item{y}{the upper y position}

\item{x}{the left x position}
}
\description{
window params
}
\keyword{internal}

