% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{with_attr}
\alias{with_attr}
\title{evaluate some code with the the given attributes on}
\usage{
with_attr(code, ...)
}
\arguments{
\item{code}{the code to evaluate}

\item{...}{passed to \code{\link{attron}} and \code{\link{attroff}}}
}
\description{
evaluate some code with the the given attributes on
}
\seealso{
Other attributes: \code{\link{COLOR_PAIR}},
  \code{\link{attroff}}, \code{\link{attron}},
  \code{\link{ncurses_attributes}}
}

