% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.r
\name{with_initscr}
\alias{with_initscr}
\title{evaluate some code inside an initialized screen}
\usage{
with_initscr(code)
}
\arguments{
\item{code}{the code to evaluate}
}
\description{
evaluate some code inside an initialized screen
}
\seealso{
Other init: \code{\link{endwin}}, \code{\link{initscr}}
}

