% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.r
\name{wmove}
\alias{wmove}
\title{perform a cursor movement in a window}
\usage{
wmove(win, y, x)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{y}{the y position (aka the row)}

\item{x}{the x position (aka the column)}
}
\value{
TRUE iff the cursor has been moved
}
\description{
may die if the position is outside the window
}
\seealso{
Other cursor: \code{\link{getyx}}, \code{\link{move}}
}

