% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refresh.r
\name{wrefresh}
\alias{wrefresh}
\title{refresh a window}
\usage{
wrefresh(win)
}
\arguments{
\item{win}{a window object as an external pointer}
}
\description{
refresh a window
}
\seealso{
Other output: \code{\link{clear}}, \code{\link{mvprintw}},
  \code{\link{printw}}, \code{\link{refresh}}
}

