% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{wsyncup}
\alias{wsyncup}
\title{touches all locations in ancestors of  win that  are changed in win.}
\usage{
wsyncup(win)
}
\arguments{
\item{win}{a window object as an external pointer}
}
\description{
touches all locations in ancestors of  win that  are changed in win.
}
\seealso{
Other window: \code{\link{border}}, \code{\link{box}},
  \code{\link{delwin}}, \code{\link{derwin}},
  \code{\link{dupwin}}, \code{\link{mvderwin}},
  \code{\link{mvwin}}, \code{\link{newwin}},
  \code{\link{subwin}}, \code{\link{syncok}}
}

