% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attributes.r
\name{attroff}
\alias{attroff}
\title{turns off the given attributes without affecting any others.}
\usage{
attroff(...)
}
\arguments{
\item{...}{any valid attributes}
}
\description{
turns off the given attributes without affecting any others.
}
\seealso{
Other attributes: \code{\link{COLOR_PAIR}},
  \code{\link{attributes}}, \code{\link{attron}},
  \code{\link{with_attr}}
}

