% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.r
\name{border}
\alias{border}
\title{draw a box around the edges of a window}
\usage{
border(win = stdscr(), ls = 0, rs = 0, ts = 0, bs = 0, tl = 0,
  tr = 0, bl = 0, br = 0)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{ls}{left side}

\item{rs}{right side}

\item{ts}{top side}

\item{bs}{bottom side}

\item{tl}{top left-hand corner}

\item{tr}{top right-hand corner}

\item{bl}{bottom left-hand corner}

\item{br}{bottom right-hand corner}
}
\description{
draw a box around the edges of a window
}
\seealso{
Other window: \code{\link{box}}, \code{\link{delwin}},
  \code{\link{derwin}}, \code{\link{dupwin}},
  \code{\link{mvderwin}}, \code{\link{mvwin}},
  \code{\link{newwin}}, \code{\link{stdscr}},
  \code{\link{subwin}}, \code{\link{syncok}},
  \code{\link{wsyncup}}
}

