% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/border.r
\name{box}
\alias{box}
\title{draw a box around a window}
\usage{
box(win = stdscr())
}
\arguments{
\item{win}{a window object as an external pointer}
}
\description{
draw a box around a window
}
\seealso{
Other border: \code{\link{border}}, \code{\link{hline}},
  \code{\link{vline}}, \code{\link{whline}},
  \code{\link{wvline}}
}

