% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{can_change_color}
\alias{can_change_color}
\title{does the  terminal  supports  colors and can change their definitions}
\usage{
can_change_color()
}
\value{
TRUE iff only it can
}
\description{
does the  terminal  supports  colors and can change their definitions
}
\seealso{
Other color: \code{\link{COLORS}},
  \code{\link{COLOR_PAIR}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{start_color}}
}

