% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.r
\name{COLOR_PAIR}
\alias{COLOR_PAIR}
\title{convert the pair to a video attribute}
\usage{
COLOR_PAIR(pair)
}
\arguments{
\item{pair}{the pair number}
}
\description{
convert the pair to a video attribute
}
\seealso{
Other attributes: \code{\link{attributes}},
  \code{\link{attroff}}, \code{\link{attron}},
  \code{\link{with_attr}}

Other color: \code{\link{COLORS}},
  \code{\link{can_change_color}},
  \code{\link{color_content}}, \code{\link{color_off}},
  \code{\link{color_on}}, \code{\link{has_colors}},
  \code{\link{init_pair}}, \code{\link{pair_content}},
  \code{\link{start_color}}
}

