% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/variables.r
\name{COLORS}
\alias{COLORS}
\title{number of colors which the terminal can support.}
\usage{
COLORS()
}
\description{
requires initializing curses
}
\seealso{
Other variables: \code{\link{COLOR_PAIRS}},
  \code{\link{COLS}}, \code{\link{ESCDELAY}},
  \code{\link{LINES}}, \code{\link{TABSIZE}},
  \code{\link{curscr}}, \code{\link{newscr}},
  \code{\link{stdscr}}
}

