% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{COLS}
\alias{COLS}
\title{the  width of the screen in characters.}
\usage{
COLS()
}
\value{
the number of columns
}
\description{
the  width of the screen in characters.
}
\seealso{
Other coords: \code{\link{LINES}}, \code{\link{getmaxyx}}
}

