% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_hello}
\alias{example_hello}
\title{hello world example}
\usage{
example_hello(sleep = 5)
}
\arguments{
\item{sleep}{sleep time before exiting}
}
\description{
hello world example
}
\seealso{
Other examples: \code{\link{example_attributes}},
  \code{\link{example_attron}}, \code{\link{example_box}},
  \code{\link{example_change_color}},
  \code{\link{example_color_set}},
  \code{\link{example_color}}, \code{\link{example_getch}},
  \code{\link{example_getnstr}},
  \code{\link{example_mouse_debug}},
  \code{\link{example_mouse}},
  \code{\link{example_mvprintw}},
  \code{\link{example_window_border}},
  \code{\link{example_window}}
}

