% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/examples.r
\name{example_mvprintw}
\alias{example_mvprintw}
\title{mvprintw example}
\usage{
example_mvprintw(sleep = 5, msg = "Just a string")
}
\arguments{
\item{sleep}{sleep time before exiting}

\item{msg}{the message to print in the center of the window}
}
\description{
mvprintw example
}
\seealso{
Other examples: \code{\link{example_attributes}},
  \code{\link{example_attron}}, \code{\link{example_box}},
  \code{\link{example_getch}},
  \code{\link{example_getnstr}},
  \code{\link{example_hello}},
  \code{\link{example_window_border}},
  \code{\link{example_window}}
}

