% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coords.r
\name{getmaxyx2}
\alias{getmaxyx2}
\title{get the current beginning coordinates and size of the default window.}
\usage{
getmaxyx2()
}
\value{
a named list of y, x (or #rows, #cols)
}
\description{
get the current beginning coordinates and size of the default window.
}
\seealso{
Other coords: \code{\link{COLS}}, \code{\link{LINES}},
  \code{\link{getmaxyx}}
}

