% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/move.r
\name{move}
\alias{move}
\title{perform a cursor movement}
\usage{
move(y, x, win = stdscr())
}
\arguments{
\item{y}{the y position (aka the row)}

\item{x}{the x position (aka the column)}

\item{win}{a window object as an external pointer}
}
\description{
may die if the position is outside the window
}
\seealso{
Other cursor: \code{\link{getyx}}, \code{\link{wmove}}
}

