% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getch.r
\name{mvwgetch}
\alias{mvwgetch}
\title{wait for user input}
\usage{
mvwgetch(win, y, x)
}
\value{
the keycode
}
\description{
wait for user input
}
\seealso{
\code{\link{KEYS}}

Other getch: \code{\link{getch}}, \code{\link{has_key}}
}

