% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{newwin}
\alias{newwin}
\title{create a new window}
\usage{
newwin(nlines = 0, ncols = 0, y = 0, x = 0)
}
\arguments{
\item{nlines}{the number of lines of the window}

\item{ncols}{the number of columns of the window}

\item{y}{the upper y position}

\item{x}{the left x position}
}
\value{
a new window as an external pointer
}
\description{
with defaults create a full-screen window
}
\details{
If either nlines or ncols is zero, they default to
     LINES - begin_y and
     COLS - begin_x.
}
\seealso{
Other window: \code{\link{box}}, \code{\link{delwin}},
  \code{\link{derwin}}, \code{\link{dupwin}},
  \code{\link{mvwin}}, \code{\link{stdscr}},
  \code{\link{subwin}}
}

