% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inopts.r
\name{raw}
\alias{raw}
\title{place the terminal into raw mode.}
\usage{
raw(bf = TRUE)
}
\arguments{
\item{bf}{whether to enable or disable}
}
\description{
Raw mode is similar to cbreak mode, in  that
characters typed are immediately passed through to the us-
  er program.  The differences are that in raw mode, the in-
  terrupt,  quit,  suspend,  and flow control characters are
all passed through uninterpreted, instead of generating  a
signal.   The  behavior  of the BREAK key depends on other
bits in the tty driver that are not set by curses.
}
\seealso{
noraw

Other inopts: \code{\link{cbreak}}, \code{\link{echo}},
  \code{\link{halfdelay}}, \code{\link{intrflush}},
  \code{\link{keypad}}, \code{\link{meta}},
  \code{\link{nocbreak}}, \code{\link{noecho}},
  \code{\link{noqiflush}}, \code{\link{noraw}},
  \code{\link{qiflush}}, \code{\link{timeout}}
}

