% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/windows.r
\name{stdscr}
\alias{stdscr}
\title{get the stdscr window as am external pointer}
\usage{
stdscr()
}
\value{
a new window as an external pointer
}
\description{
get the stdscr window as am external pointer
}
\seealso{
Other window: \code{\link{box}}, \code{\link{delwin}},
  \code{\link{mvwin}}, \code{\link{newwin}}
}

