% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addch.r
\name{waddch}
\alias{waddch}
\title{put a character
put the
      character ch into the given window at its  current  window
      position,  which  is then advanced.  They are analogous to
      putchar in stdio(3).  If the advance is at the right  mar-
      gin:}
\usage{
waddch(win, ch)
}
\arguments{
\item{win}{a window object as an external pointer}

\item{ch}{a character to print}
}
\description{
put a character
put the
      character ch into the given window at its  current  window
      position,  which  is then advanced.  They are analogous to
      putchar in stdio(3).  If the advance is at the right  mar-
      gin:
}
\seealso{
Other addch: \code{\link{addch}}, \code{\link{echochar}},
  \code{\link{mvaddch}}, \code{\link{mvwaddch}},
  \code{\link{wechochar}}
}

